Refactored the way parsing arguments work.
Now, we have a main parser created at the process_img.py file.
Then, we add a subparser for every existing module, these subparsers will take care of parsing that module's arguments.
Each module also has a 'command' argument which is set to that module's name and used as a identifier to know which module was called.